requirements.txt
easyocr==1.7.2
filelock==3.16.1
fsspec==2024.12.0
imageio==2.36.1
Jinja2==3.1.5
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.1
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pyclipper==1.3.0.post6
PyMuPDF==1.25.1
python-bidi==0.6.3
PyYAML==6.0.2
scikit-image==0.25.0
scipy==1.14.1
setuptools==75.6.0
shapely==2.0.6
sympy==1.13.1
tifffile==2024.12.12
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
